Introduce caching in poco injection
