+++ Packages/com.deltation.di-framework/Assets/DELTation/DIFramework/Runtime/Resolution/PocoInjection.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 822674bd6e2e41499d8c080ce5c7f920
+timeCreated: 1619121256
